+++ mr/migrator/browser/helper.py
@@ -142,7 +142,6 @@ class MigratorRun(group.GroupForm, form.Form):
             section, key = key.split('-')
             overrides.setdefault(section, {})[key] = value
 
-#        import ipdb; ipdb.set_trace()
         logger.info("Start importing profile: " + data['config'])
         Transmogrifier(self.context)(data['config'], **overrides)
         logger.info("Stop importing profile: " + data['config'])
@@ -153,16 +152,10 @@ class MigratorConfigurations(object):
     def __call__(self, context):
         terms = []
         for conf_id in configuration_registry.listConfigurationIds():
-            print conf_id
-            conf_file = _load_config(conf_id)
-            for section_id in conf_file.keys():
-                section = conf_file[section_id]
-                if section.get('blueprint', '') ==\
-                        'plone.app.transmogrifier.atschemaupdater':
-                    conf = configuration_registry.getConfiguration(conf_id)
-                    terms.append(SimpleVocabulary.createTerm(
-                        conf_id, conf_id, conf['title']))
-                    break
+            conf = configuration_registry.getConfiguration(conf_id)
+            terms.append(SimpleVocabulary.createTerm(
+                conf_id, conf_id, conf['title'])
+            )
         return SimpleVocabulary(terms)